Move components to storybook folder
