+++ src/components/Canvas/CanvasInner.default.tsx
@@ -1,4 +1,3 @@
-import styled from 'styled-components'
 import { IConfig, IOnCanvasClick } from '../../'
 
 export interface ICanvasInnerDefaultProps {
@@ -11,11 +10,3 @@ export interface ICanvasInnerDefaultProps {
   onDrop: (e: React.DragEvent<HTMLDivElement>) => void
   onDragOver: (e: React.DragEvent<HTMLDivElement>) => void
 }
-
-export const CanvasInnerDefault = styled.div<ICanvasInnerDefaultProps>`
-  position: relative;
-  outline: 1px dashed rgba(0,0,0,0.1);
-  width: 10000px;
-  height: 10000px;
-  cursor: move;
-` as any
