+++ src/components/Canvas/CanvasOuter.default.tsx
@@ -1,4 +1,3 @@
-import styled from 'styled-components'
 import { IConfig } from '../../types'
 
 export interface ICanvasOuterDefaultProps {
@@ -7,15 +6,3 @@ export interface ICanvasOuterDefaultProps {
   children: any
   ref?: React.Ref<any>
 }
-
-export const CanvasOuterDefault = styled.div<ICanvasOuterDefaultProps>`
-  position: relative;
-  background-size: 20px 20px;
-  background-color: rgba(0,0,0,0.08);
-  background-image:
-    linear-gradient(90deg,hsla(0,0%,100%,.2) 1px,transparent 0),
-    linear-gradient(180deg,hsla(0,0%,100%,.2) 1px,transparent 0);
-  width: 100%;
-  overflow: hidden;
-  cursor: not-allowed;
-` as any
