+++ src/components/Port/Port.default.tsx
@@ -1,5 +1,3 @@
-import * as React from 'react'
-import styled from 'styled-components'
 import { IConfig, IPort } from '../../'
 
 export interface IPortDefaultProps {
@@ -11,33 +9,3 @@ export interface IPortDefaultProps {
   isLinkSelected: boolean
   isLinkHovered: boolean
 }
-
-const PortDefaultOuter = styled.div`
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
-  background: white;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  &:hover > div {
-    background: cornflowerblue;
-  }
-`
-
-const PortDefaultInner = styled.div<{ active: boolean }>`
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  background:  ${(props) => props.active ? 'cornflowerblue' : 'grey' };
-  cursor: pointer;
-`
-
-export const PortDefault = ({ isLinkSelected, isLinkHovered, config, className }: IPortDefaultProps) => (
-  <PortDefaultOuter className={className}>
-    <PortDefaultInner
-      active={!config.readonly && (isLinkSelected || isLinkHovered)}
-    />
-  </PortDefaultOuter>
-)
