+++ stories/components/NodeInner.default.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import styled from 'styled-components'
+import { INodeInnerDefaultProps } from '../../src'
+
+const Outer = styled.div`
+  padding: 40px 30px;
+`
+
+export const NodeInnerDefault = ({ node,className }: INodeInnerDefaultProps) => {
+  return (
+    <Outer className={className}>
+      <div>Type: {node.type}</div>
+    </Outer>
+  )
+}
