+++ stories/components/index.ts
@@ -1,6 +1,11 @@
+export * from './CanvasInner.default'
+export * from './CanvasOuter.default'
 export * from './Code'
 export * from './Content'
 export * from './Message'
+export * from './Node.default'
+export * from './NodeInner.default'
 export * from './Page'
+export * from './Port.default'
 export * from './Sidebar'
 export * from './SidebarItem'
